+++ coordinator/statement_executor.go
@@ -63,10 +63,10 @@ func (e *StatementExecutor) ExecuteStatement(stmt influxql.Statement, ctx *influ
 			messages = append(messages, influxql.ReadOnlyWarning(stmt.String()))
 		}
 		if stmt.IfNotExists {
-			ctx.Log.Println("WARNING: IF NOT EXISTS is deprecated as of v0.13.0 and will be removed in v1.0")
+			ctx.Log.Println("WARNING: IF NOT EXISTS is deprecated as of v0.13.0 and will be removed in a future release")
 			messages = append(messages, &influxql.Message{
 				Level: influxql.WarningLevel,
-				Text:  "IF NOT EXISTS is deprecated as of v0.13.0 and will be removed in v1.0",
+				Text:  "IF NOT EXISTS is deprecated as of v0.13.0 and will be removed in a future release",
 			})
 		}
 		err = e.executeCreateDatabaseStatement(stmt)
@@ -95,6 +95,13 @@ func (e *StatementExecutor) ExecuteStatement(stmt influxql.Statement, ctx *influ
 	case *influxql.DropDatabaseStatement:
 		if ctx.ReadOnly {
 			messages = append(messages, influxql.ReadOnlyWarning(stmt.String()))
+		}
+		if stmt.IfExists {
+			ctx.Log.Println("WARNING: IF EXISTS is deprecated as of v0.13.0 and will be removed in a future release")
+			messages = append(messages, &influxql.Message{
+				Level: influxql.WarningLevel,
+				Text:  "IF EXISTS is deprecated as of v0.13.0 and will be removed in a future release",
+			})
 		}
 		err = e.executeDropDatabaseStatement(stmt)
 	case *influxql.DropMeasurementStatement: